Update TheButton.lua
I didn't understand how the original script worked and I also wanted my button presses to toggle the light. So I modded it so that if a light is turned on it will turn off when you press the button and vice versa. I have very poor knowledge in lua script and python. So please feel free to improve on my script.